This is our website for crisis management platform
This is a website which I created on SITE-24 Hackathon. In the website their is a ML Model for fire detection using camera's IP address. We also put a Weather report API in this website. Our future plan for this website is to add an alert system for fire detection and a rainfall prediction model.